project_02_tool_agent/requirements.txt
langchain>=0.3
langchain-community>=0.3
langchain-core>=0.3
langgraph>=0.2
duckduckgo-search>=6.0
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.40
python-dotenv>=1.0
loguru>=0.7
httpx>=0.27
pydantic>=2.0
